+++ src/consensus/pbft/pbft_globals.h
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
-#include "consensus/pbft/libbyz/ITimer.h"
-#include "consensus/pbft/libbyz/Statistics.h"
+#include "consensus/pbft/libbyz/itimer.h"
+#include "consensus/pbft/libbyz/statistics.h"
 #include "node/node_to_node.h"
 
 #include <signal.h>
